requirements.txt
# ==========================================
# ALGO TRADER v3.0 - REQUIREMENTS
# ==========================================

# === CORE DEPENDENCIES ===
MetaTrader5>=5.0.45
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.0
PyYAML>=5.4.0
pytz>=2021.1

# === AI/LLM ===
openai>=1.0.0

# === DATABASE ===
SQLAlchemy>=1.4.0
alembic>=1.7.0

# === WEB INTERFACE ===
streamlit>=1.10.0

# === TECHNICAL ANALYSIS ===
ta>=0.10.0  # Technical Analysis library
scipy>=1.7.0
statsmodels>=0.12.0

# Note: TA-Lib requires separate binary installation
# Download from: https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
# Then: pip install TA_Lib‑0.4.24‑cp310‑cp310‑win_amd64.whl

# === MACHINE LEARNING ===
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0
optuna>=2.10.0  # Hyperparameter optimization

# === BACKTESTING ===
backtrader>=1.9.0
vectorbt>=0.24.0

# === VISUALIZATION ===
matplotlib>=3.4.0
plotly>=5.3.0
seaborn>=0.11.0

# === PERFORMANCE ===
numba>=0.54.0  # JIT compilation
cython>=0.29.0  # C-extensions

# === MONITORING ===
psutil>=5.8.0  # System monitoring
watchdog>=2.1.0  # File system monitoring

# === UTILITIES ===
aiohttp>=3.8.0  # Async HTTP
websocket-client>=1.2.0  # WebSocket support
python-dateutil>=2.8.0
colorlog>=6.6.0  # Colored logging
tqdm>=4.62.0  # Progress bars
schedule>=1.1.0  # Task scheduling

# === CRYPTO EXCHANGES (Optional) ===
ccxt>=2.5.0  # Cryptocurrency exchange library
python-binance>=1.0.0  # Binance API

# === DEVELOPMENT ===
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-mock>=3.7.0
black>=22.0.0  # Code formatter
flake8>=4.0.0  # Linter
mypy>=0.930  # Type checker

# === TELEGRAM ===
# Note: python-telegram-bot is optional, using requests instead
# python-telegram-bot>=13.0

# === DATA SOURCES ===
yfinance>=0.1.70  # Yahoo Finance (backup data source)
alpha-vantage>=2.3.0  # Alpha Vantage API (optional)

# === ADDITIONAL ML (Optional) ===
# tensorflow>=2.8.0  # For LSTM models
# keras>=2.8.0
# torch>=1.10.0  # PyTorch alternative
# lightgbm>=3.3.0  # Alternative to XGBoost

# === VERSION COMPATIBILITY ===
# Python >= 3.10 required
# Windows 10/11 64-bit recommended
# MetaTrader 5 build 3550+